LuwiChess homepage is now complete!
This fixes the product page and its offcanvas shopping cart to use javascript arrays to display the products.
nav.css was created to not use the same css in multiple files, since the navigation should be the same on every page.
The opening hours in the footer were moved to a table on all pages.
The mobile view on the index page now shows the large chess image above the "Become a Chess Pro!" text, instead of disappearing.
All comments are now in the same language (Swedish).
Test files were removed.
The hamburger menu is now white on all pages.