This changes the html for total amount in the shopping cart.
This changes the html for total amount in the shopping cart, which was previously set to $999.99 to help me see that the javascript code worked. The js-code immediately overwrites the amount to zero dollars, but it is unneccessary to keep that control number, even if the users will not be able to see it (unless they inhibit javascript from running in their browser).